review(test): drop (int)chainLength casts at OfChainLength call sites where source is ulong
BlockTreeTests / EraETestModule / BlockDownloaderTests.SyncPeerMock.BuildTree:
source variable is already ulong; the (int) cast was forcing the int overload
unnecessarily. Use the existing ulong OfChainLength overload directly.

(ForwardHeaderProviderTests.cs:551 retains its (int) cast since `chainLength`
is long there, not ulong.)

+++ src/Nethermind/Nethermind.Blockchain.Test/BlockTreeTests.cs
@@ -1682,7 +1682,7 @@ public class BlockTreeTests
             PivotNumber = pivotNumber
         };
 
-        BlockTreeBuilder treeBuilder = Build.A.BlockTree().OfChainLength((int)head + 1);
+        BlockTreeBuilder treeBuilder = Build.A.BlockTree().OfChainLength(head + 1);
 
         BlockTree loadedTree = Build.A.BlockTree()
             .WithoutSettingHead
