+++ src/Nethermind/Nethermind.Synchronization.Test/BlockDownloaderTests.cs
@@ -1064,7 +1064,7 @@ public partial class BlockDownloaderTests
                 builder = builder.WithTransactions(_receiptStorage);
             }
 
-            builder = builder.OfChainLength((int)chainLength, 0, 0, _withWithdrawals);
+            builder = builder.OfChainLength(chainLength, 0, 0, _withWithdrawals);
             BlockTree = builder.TestObject;
 
             HeadNumber = BlockTree.Head!.Number;
